policy-service/requirements.txt
# policy-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
redis
python-dotenv